Change logic in arm1pps command
As originally written, the arm1pps command would hang waiting for
a state
change that never happened.  This changes to logic to work with what we
observed in hardware, BUT what we observed in hardware looks like
triggering is
occurring on the falling edge of 1 pps rather than the rising edge
of 1 pps
even though the design looks like it is supposed to trigger on the
rising edge.
Very strange.