Add file vlbi_dsp_core.mdl
This is used to block box VLBI DBE DSP functionality including frequency
channelization and complex-to-real conversion/interpolation.  The output of
this core is designed to feed into the VSI data formatter.

This version includes the convert block updates to the twiddle general blocks
in casper_library as of git commit d9fdbd0e7ff67d44560c982815e97c610f91ba36.