Fixed crash causing typo in init scripts
The coef_load_init.m and shift_reg_init.m files used find_system with
'LookUnderMasks' set to 'on'.  This is not a correct value.  It should be 'all'
rather than 'on'.  The behavior of simulink when find_system is passed this
incorrect value is not to error out as one might expect, but rather to do
something different and (apparently) undocumented.  Apparently, passing this
incorrect value within a mask initialization script will cause the mask
initialization script to run again and again and again and so on until some
recursion limit is reached and Simulink/Matlab crashes.