Add 1 PPS generator block to ibob_phased_array_2k
In order to facilitate synchronized seeding of noise sources on different
ibobs, the noise generators are armed via a software register and then load the
new seed on the next sync pulse.  SMA uses a 100 Hz heartbeat for their sync
pulse, which makes for tight timing on the arming, so an internal 1 PPS is
generated synchronous with one or the other of the ADCs sync pulses (user
selectable).  It is this internal 1 PPS that is used to load the new seeds.

Also redrew FIR Filter blocks to get latest beamform library updates/fixes.