Convert XAUI links to use demux of 1
The XAUI links had used a demux of two, which uses a 32-bit data path.  Now
they use a demux of 1 and a 64-bit data path.  This is being done because of a
suspected (though as yet uncomfirmed) bug in the demux of 2 mode.  Using this
model might require that the XAUI blocks in the ibob_phased_array_2k model also
change to use demux of 1.  The "valid" register which shows the state of the
rx_valid output from the XAUI block noe has two significant bits instead of
one.  The least significant bit is the current value and the next bit is the
value from one cycle ago.  Since this should be toggling every cycle, the
"valid" register should always be 1 or 2.  A 0 or 3 means that something is
weird.

The XAUI blocks have been renamed to match the ibob faceplate
naming convention (which is opposite from the yellow port naming convention).
The actual ports used remain the same however, so this is only a naming
convention change; it will NOT require any recabling of existing systems since
only the names have been changed.

The XAUI blocks have also been somewhat rearranged to provide a naming
convention witch is consistent with the naming convention in the
ibob_phased_array_2k model.