Convert IPA XAUI links to use demux of 1
The XAUI links had used a demux of two, which uses a 32-bit data path.  Now
they use a demux of 1 and a 64-bit data path.  This is being done because of a
suspected (though as yet uncomfirmed) bug in the demux of 2 mode.

The XAUI blocks have been renamed to match the ibob faceplate
naming convention (which is opposite from the yellow port naming convention).
The actual ports used remain the same however, so this is only a naming
convention change; it will NOT require any recabling of existing systems since
only the names have been changed.