Fix adder tree bit growth handling
The adder tree would chop off bits prematurely from the intermediate sums.
This fix maintains full precision throughout the adder tree, then requantizes
to 8 bits by rounding off (using round to even) the lower one or two bits and
dropping the most significant bit (with saturation) when rounding off one bit.

The selection of how many bits to round off is controlled by the least
significant bit of the newly added "sum_shift" register.  When it is set to 0
(the default), it rounds off 1 bit.  This is suitable for incoherent noise
dominated inputs (adding 4 independent random values together has an expected
growth factor of sqrt(4) == 2).  When it is set to 1, it rounds off 2 bits.
This is suitable for signals that are not incoherent noise dominated (e.g. tone
or coherent noise) since the expected growth would then be 4.