Modified correlator_core to use the special 2-bit encoding.
Note that it now requires an additional input port which is a
dynamically set threshold for quantization. I'm still having trouble
with timing, in this case the quantizer relational blocks are
failing their constraints.

Will look into this further and once a fully compiled design has been
tested I will tag and add a detailed explanation of how to use the
correlator.