Add xaui aligner and monitor to dbe_xaui_corr
This resulted in changing the meaning of "delay_adj", expanding the width and
scope of "xaui_rst", and creating several new registers for monitoring the
status of the xaui links.

The "onepps_sync" block now outputs both 1 PPS and 1024 PPS and the 1024 PPS
signal can be output over sma1 by setting "sma1pps_sel" to 2.

The sync passed into the vsi_dsp block is now the 1 PPS sync pulse output from
one of the aligned out-of-band signals.

This design will only work with ibob_phased_array_2k designs that output 1024
PPS on the least significant bit of the OOB signals and 1 PPS on the next bit.
It will NOT work if the upstream ibob_phased_array_2k systems do NOT output
these sync signals.