64-bit XAUI with single-baseline correlator now successfully compiles.
Timing issues were corrected using a combination of modifications and
additions to the cfa_library/cross_correlator block and adding latency
to the correlator_core and dbe_xaui_corr models. I will add the bitstream
after I back-port the correlator to the 32-bit XAUI interface for testing
purposes. I will then tag these with a link to the version of cfa_library
and mlib_devel used to compile them.