Add phase compensation to delay_core block
Because the delay_core's fine delay (i.e. 0, 1, 2, or 3 ADC samples) precedes
the complex +/- F2/4 LO used in the real-to-complex conversion, a phase offset
is introduces when the fine delay it 1, 2, or 3.  This commit adds a
"phase_compensation" block (to the sub_delay block) to compensate for this
automatically.  The phase compensation block also compensates for the 180
degree phase shift caused by negating the coefficients of the qfdfir (since -1
can be represented in Fix_16_15, but +1 cannot).

This commit also includes the delay_core_test.mdl file that was used to test
the delay_core block.