Added a complex correlator core for use in the BEE2 calibrator.
This single-baseline 2-bit complex correlator is intended as a block box
core to be used to populate the BEE2 calibration correlator. Inputs are
two parallel complex samples from two separate antennas and the output
is the 16-lag complex correlation (given a dynamically adjustable integ-
gration time, also an input).

After this is fully tested I will generate the netlist and include the
'cfa_library' commit used. Then I will create a new calibration correlator
model file with XAUI monitoring and control similar to those on the DBE
and populate it with 7 (possibly at some point more) of the black box
netlist core.

Software for control and documentation coming soon.