Fix sync period bug in vlbi_dsp_core
The wideband real FFT used in vlbi_dsp_core has a reorder block that is of
order 3.  This means that the sync period for that FFT block must be a multiple
of 3.  The sync period being passed in to the FFT block had been 256e6, which
does not have a factor of 3.  This resulted in a glitch in the FFT output every
sync pulse.

To fix this, the incoming sync is gated to allow only every third sync pulse
through.  This provides the FFT with a sync period that has a factor of 3.  The
sync_out from the FFT has the same period as the sync_in.  This 3x period sync
pulse is used to synchronize the regeneration of a 1 PPS signal, which the
downstream logic (and data recorders) require.