Add Walsh function look-up tables
Walsh look-up tables are stored in two shared BRAMs: walsh/table/90 for the
Walsh table for 90 degree phase switching and walsh/table/180 for the Walsh
table for 180 degree phase switching.  Each 32-bit word from the table is a row
of a 32 row by 32 column Walsh table.  Each row contains the phase switch
states for 32 inputs.  Each column of the Walsh table represents a phase
switching Walsh function for one input.  The Walsh function (i.e. column) used
for each ADC input is selected via the walsh/colsel register.  Each byte of
this 32-bit register specifies the column to be used for the corresponding
input (LSB is in0, MSB is in3).

The 90 and 180 degree phase state bits for a given input are concatenated to
form a 2 bit "phase quadrant" value:

00 =   0 degrees
01 = +90 degrees
10 = 180 degrees
11 = -90 degrees

This phase quadrant is *subtracted* from the phase_offset given in the input's
"phaseN" register.

The phase state info is sent to the calibration correlator over the XAUI OOB
signals.  The demux to 64/8 data/oob bits is a bit tricky wince we want to send
6 synchronous OOB signals and 1 asynchronous OOB signal (sync0).  The
calibration correlator will need to treat the re-mux of the data and OOB
signals carefully.