Tentatively added support for SMA Walsh generation to IPA.
Added a 'walsh/sync_sel' register to choose between the CARMA
sync period (default) and the SMA sync period, which is called
the heartbeat and is not an integer multiple of the clock rate.

And I added a 'hb_sync' block (modeled after the '1pps_sync' block)
which sync's to an external "start of Walsh frame" signal (currently
coming from ADC0/sync0 but will likely be changed to a gpio) and
generates its own SOWF and heartbeat pulses. Like '1pps_sync' it can
be armed at any time and will likely need to be re-armed every few
hours.

Now to compiling...