Add sideband separation to complex_cc_core
This requires two more inputs, in0_w and in1_w, to indicate the Walsh I/Q state
for each input.  It also changes the output data ports to be usb_data_re and
usb_data_im (for the upper sideband data) and adds two more output ports,
lsb_data_re and lsb_data_im (for lower sideband data).

The names of the input ports were also changed to avoid overloading the 'I' and
'Q' terminology.

The test inputs are also changed to induce an 8 sample wide auto-correlation
"pulse" that produces a triangle wave in the output.

Several WaveScope blocks have also been added to the design to help with
visualizing the flow of data through the system.

This version does not force the integration time to be an integer multiple of
the sync_in period.