Support new sync-over-oob convention in BCC
The new sync-over-oob convention is...

oob={w3[0],w2[0],w1[0],w0[0],pps1,sowf,sync0,pps1024}

This puts the sync signals in "period order" in that the signals with larger
periods are in more significant positions, i.e.  bit3=slowest, bit2=slow,
bit1=fast, bit0=fastest.   The pps1024 is in the least significant bit of oob
as expected/required by the receiving xaui aligner.

In the bee2_complex_corr, the signals are unpacked according to the above
expression and the four sync signals are fed into a mux whose select line is
driven by the lowest two bits of the "syncsel" register.  This allows the user
to choose any of the four sync signals as the correlator sync signal.  Once
complex_cc_core's integration timing is changed to integrate in multiples of
the slected sync period, the integ_time register will be in units of the
selected sync pulse and the actual integration time used will be rounded up to
the next multiple of 16.  I'm not sure if we want to enforce each
sub-integration to be a multiple of 16 cycles or not.  I think we do or we risk
mis-weighting the lags (i.e. some lags will sum more samples than other lags).

syncsel = 0 ==> use pps1024 (1024    PPS)
syncsel = 1 ==> use sync0   ( 100    PPS heartbeat)
syncsel = 2 ==> use sowf    (   1.56 PPS SOWF)
syncsel = 3 ==> use pps1    (   1    PPS)