Make complex_cc_core integrate in units of sync periods
The main change is to make complex_cc_core integrate in units of sync periods.
This is complicated by the fact that the SMA sync pulse is not commensurate
with the integration sample rate.  The correlator could simply be reset every N
sync pulses regardless of where the correlator is in its integration cycle, but
it turns out that the simple_bram_vaccs do not take kindly to starting a new
accumulation in the the middle of an ongoing accumulation.  When that happens,
the simple_bram_vaccs output in non-standard order.  This would result in
"corrupted" integrations.  Given that the SMA sync pulse is not commensurate
with the integration sample rate, this would result in large amounts of invalid
data.

The solution used here is to assert a "new_acc_req" signal on the Nth sync
pulse.  This is ANDed with the cross correlator's "valid" output signal thereby
generating a "new_acc" pulse on the next "valid" pulse which is commensurate
with the correlator sample time.  The correlator itself is only reset when the
integ_time register changes; otherwise it is left free-running.  Changing
integ_time will therefore result in one corrupted integration approximately
15/16 of the time (i.e. there is a 1-in-16 chance that its change will reset
the correlator at the precise time the simple_bram_vaccs expect to start a new
accumulation).

The new control logic also provides two new outputs with hopefully useful
information.

Here is a summary of what was done:

1. Integration is specified in units of sync periods

2. Handle non-commensurate sync periods gracefully.  All lags in a given
   integration have the same number of samples, but some integrations may not
   have the same number samples as other integrations.

3. Provide two new outputs:

   3.1 "sample_cnt" contains the number of samples (per lag) that were summed
       during the last integration

   3.2 "subint_cnt" contains the count of sync pulses within the current
       integration (progress indicator)

4. Changing integ_time causes next sync pulse to reset cross_correlator and
   "integs" register.

5. Added several WaveScope blocks and named lots of signals to aid in
   simulating/debugging.