Move xaui alignments closer to middle of ASRs
Because bee2_complex_corr does not get any external sync signals, it can only
align the two incoming XAUI streams relative to each other rather than an
external reference.  To do this, one of the incoming sync signals is delayed
and considered to be the reference to which both streams should be aligned.
The alignment occurs by determining how far from the reference sync each
stream's sync is and then delaying each stream by that amount.  This depends on
the reference sync being delayed slightly as compared to the streams' syncs.
Ideally, the reference sync will be delayed by one half the depth of the
addressable shift register (ASR) that is used as a variable delay.  This gives
the streams the most breathing room in either direction.

One of the incoming syncs was delayed by 1 cycle to create the reference sync.
This results in many cycles of "breathing room" in one direction, but fewer
than desired cycles of breathing room in the other direction.  To fix this, the
delay was changed to 32, which is half the depth of the ASRs.