Improve complex_cc_core testbench
The complex_cc_core.mdl file includes some testbench functionality, but it has
involved a fairly simplistic test case involving the autocorrelation of a 8
cycle wide rectangular pulse.  This is great for testing some basic
functionality, but it is not so great for testing certain timing aspects of the
design.

In-chip system testing of the calibration correlator revealed that the
autocorrelation lag spectrum was not Hermitian symmetric as expected for an
autocorrelation.  I suspect this is because of slight time skew in the
integrated lags.  For example, in a given integration, lag -8 might go from T0
to T1 whereas lag +7 might go from T0+delta to T1+delta (i.e. a relative time
skew of 'delta' between lags -8 and +7).

This commit enhances the complex_cc_core testbench so that the user can easily
choose between the earlier rectangular pulse autocorrelation test or a new
random input autocorrelation test.  The new random input autocorrelation test
exposes the asymmetric autocorrelation lag spectrum bug as observed in in-chip
system testing.  To see the anti-symmetric lag spectrum, simulate the model
with the noise inputs selected, then change to the matlab prompt and run...

complex(real_cf([0:15]+353), imag_cf([0:15]+353))

This will create and show the complex autocorrelation lag spectrum from the
testbench's "to workspace" blocks.  It will be apparent that the
autocorrelation lag spectrum has a strong peak at lag 0 as expected, but lacks
Hermitian symmetry (i.e. even symmetry in the real component and odd symmetry
in the imaginary component).

This commit does NOT fix this bug; it merely adds a testbench that exposes this
bug.  A future commit will fix this bug.