Redraw the cross_correlator in complex_cc_core
The cross_correlator block was redrawn to pick up fixes from the cfa_library.
The cross_correlator block was redrawn using cfa_library tag
"cfa_library-2011-02-11".

Both the impulse autocorrelation test and the random input autocorrelation test
now show symmetric autocorrelation lag spectra.  Lag-8 has no corresponding
positive lag with which to be symmetric, so it must be excluded for a purely
symmetric autocorrelation lag spectrum.  This is inherent in an
even-number-of-taps lag correlator where the lags are all integral.  To see the
(mostly) symmetric lag spectrum, simulate the model with the noise inputs
selected, then change to the matlab prompt and run...

complex(real_cf([0:15]+346), imag_cf([0:15]+346))

This will show the complex autocorrelation lag spectrum from the
testbench's "to workspace" blocks.  The values shown correspond to lag-8 to
lag+7.  It will be apparent that the autocorrelation lag spectrum from lag-7 to
lag+7 is Hermitian symmetric (i.e. even symmetry in the real component and odd
symmetry in the imaginary component) as expected.