Add wdelay and phsw_bal ports to complex_cc_corr
The wdelay port is a UFix_9_0 input port that specifies the delay to apply to
the walsh signals to align them with the data flowing out of the
cross_correlator block.  The latency through the cross_correlator block is
nominally:

4 for the inital in1 delay
+ 2*Nlag/2 for the first Nlag/2 stages
+ 3*Nlag/2 for the second Nlag/2 stages
+ 5 for lag stage math latency
+ 16 for the stream accumulator
------------------------------------------------------
4 + 16 + 24 + 5 + 16 == 65

This calculation probably has some error, so to be safe we use an adjustable
delay that can be set as needed at runtime.  If a certain value proves
consistently good (one should) then the adjustable delay can be replaced with
a suitable fixed delay.  The phsw_bal value can be used to determine whether
the adjustable delay is good.

The phsw_bal output port indicates the phase switching balance for the most
recently completed integration.  During the integration, a counter
(accumulator, really) counts +1 for each sample where the inputs were in the
same 90 degree phase switch state and -1 for each sample where the inputs were
in different 90 degree phase switch states.  For optimum sideband separation,
the inputs should have an equal number of samples while in the same and
different 90 degree phase switching states, so ideally this output should read
zero.  If it is positive, then there that many more samples with the same 90
degree phase switch state.  If it is negative, then there that many more
samples with different 90 degree phase switch state.