Update bee2_complex_corr model with latest fixes
The bee2_complex_corr model file is updated to include the latest
complex_cc_corr netlist.  This updated netlist includes two new ports: wdelay
and phsw_bal.  The wdelay input port on all baselines is driven by a common
"wdelay" register.  The phsw_bal output ports go to baseline specific software
registers ("phsw_bal/0" through "phsw_bal/7").

A snap64 block from casper_library_scopes was added to provide access to the
2-bit samples coming into the correlator.  The snap64 block's link was disabled
and the bram names changed to something more meaningful and less likely to
cause confusion (one had been named "bram_lsb", but it certainly is not "lower
sideband").  The samples for inputs 0-3 are stored in a shared BRAM named
"snap/in0123".  The samples for inputs 4-7 are stored in a shared BRAM named
"snap/in4567".  The samples for the lowest numbered input for a BRAM are stored
in the most significant byte.  The samples for the highest numbered input for a
BRAM are stored in the least significant byte.  Since these are 2-bit samples,
each byte obviously contains four samples.  The two most significant bits of
each byte are the first sample of the four.  The snap BRAMs are 2^11 (2048)
words deep.  The snapshot can be triggered via software or armed via software
to be triggered on the next rising edge of the selected sync signal.

Finally, an uptime generator with software register were added.  The uptime
generator simply counts seconds since the FPGA started up.  There is no way to
reset it without restarting the FPGA (other than waiting more than 4 billion
seconds!).  There is no way to synchronize it with an external 1 PPS.