Change wdelay implementation to improve timing
wdelay had been delays by a 128 deep adjustable shift register.  This caused
timing problems.  To deal with this, 31 cycles of latency were moved before the
adjustable shift register, the adjustable shift register was shortened to 64
deep, and one cycle of latency was added after the adjustable shift register.

The best guess at this point is that a value of 32 at the wdelay port will be
ideal, but this is subject to change based on reality.