Add ibob_phased_array_2k_2011_Feb_17_0251.mcs
This commit will be tagged as "ibob_phased_array_2k_2011_Feb_17_0251".
Note: this was compiled with tag vlbi-devel-2011-02-16 of the RAL fork of the
mlib_devel library and tag cfa_library-2011-02-16 of cfa_library.

The relevant change to mlib_devel was to the XAUI clock IO standard (changed
from LVDS_25 to LVDS_25_DT).

The relevant change to cfa_library was the addition of a reset port to the
2-bit quantizer blocks.