Increase integration time limit
The BRAM vector accumulators were 32 bits wide.  This limits the number of
product samples that can be accumulated.  Given that each two bit sample
represents one of the four values [-3, -1, +1, +3] and assuming that
thresholding is set such that the probabilities of each value are [1/6, 1/3,
1/3, 1/6], respectively, then each sample of the lag-0 autocorrelation takes on a
value of [9, 1, 1, 9] with the same probabilities.  This product distribution
has a mean of 11/3 and a variance of 128/9.  Integrating this for one second
second (i.e. 1024e6 samples) will have a mean of 11/3*1024e6 and a variance of
128/8*1024e6.  This integrated mean is more 3.75e9, which is well over the
maximum value supported by a signed 32-bit value (2.1475e9), meaning that the
maximum integration time is less than one second!

This integration time limit has been increased through two changes:

1) Widen the BRAM vector accumulators to 36 bits.  This will allow for up to 9
second integrations.

2) Drop LSb before BRAM vector accumulators.  Since all the input samples are
odd, their products are also odd.  Since we always add an even number of odd
products together in each sub-integration before the BRAM vector accumulators,
the values going into the BRAM vector accumulators are always even.  There is
no value in carrying this extra '0' in the least significant bit, so it can
simply be dropped.  This doubles the maximum integration time.  Note that this
change could be implemented further upstream (e.g. just after the complex
multipliers).  If it is ever implemented further upstream, be sure to
un-implement it here.

These two changes together should allow for integrations as long as 18 seconds
without overflowing the accumulators.