Add ext_period and ext_cnt registers on SOWF/1PPS
These can be used to verify that the external SOWF and 1PPS signals are present
and have the expected periods in terms of FPGA clock cycles.

Also ensure that these counters/registers (and the diff counter) are only
sensitive the the rising edge of the external signal, not the level.