Make Walsh tables use unrolled Walsh functions
Previously the two Walsh tables were addressed using different bits from a 10
bit counter.  The upper 5 bits were used as the address to the 90 degree table
and the lower 5 bits were used as the address to the 180 degree table.  This
allowed storing nested Walsh function more compactly, but was a problem because
the SOWF signal reset the counter every 64 pulses, so only the first two
locations of the 90 degree table were ever output.

The counter is now used to address both Walsh tables directly.  This means that
both Walsh tables must store their Walsh functions in "unrolled" form.  For
CARMA, this means that each table stores 1024 values.  For SMA, this means that
each table stores 64 values.