Add ibob_phased_array_2k_2011_Feb_21_1459.mcs
This commit will be tagged as "ibob_phased_array_2k_2011_Feb_21_1459".
Note: this was compiled with tag vlbi-devel-2011-02-16 of the RAL fork of the
mlib_devel library (uses LVDS_25_DT for XAUI clocks) and tag
cfa_library-2011-02-16 of cfa_library.

This bitstream met timing, but took 5.75 hours to build.  Looking at the
verbose timing report (created manually with trce), it seems like some changes
that might make things less tight would be to change the delays in the phase
rotators to "use behavioral HDL" and slightly increase the latency in front if
the phase rotators' multipliers.  Other possible changes would be to change the
latency in the delay core to ease up its timing a little.