Add walsh/sign90 register to IPA design
The LSb of the walsh/sign90 register specifies the sign of 90 degree phase
switch in the upper sideband.

When set to 0:

    The analog 90 degree phase switching is assumed to be a +90 degree phase
    rotation in the USB and a -90 degree phase rotation in the LSB.  The
    complex signal is phase rotated by -90 degrees, which rotates the USB to 0
    degrees and the LSB to 180 degrees.

When set to 1:

    The analog 90 degree phase switching is assumed to be a -90 degree phase
    rotation in the USB and a +90 degree phase rotation in the LSB.  The
    complex signal is phase rotated by +90 degrees, which rotates the USB to 0
    degrees and the LSB to 180 degrees.

Depending on which Nyquist band is sampled and the setting of the "sbsel"
register, the apparent sign of the 90 degree phase rotation in the upper
sideband may be different from the locally defined convention.