Added LUTs for phase and delay tracking.
The look-up tables (LUTs) output a new value after every 0/180 Walsh cycle.  If
the LSb of lut/enable is 1, the outputs of the delay and phase luts are added
to the corresponding delay and phase "constant" to produce the net delay and
phase to be applied.  If the LSb of lut/enable is 0, then the outputs of the
LUTs are ignored leaving just the "constant" delay and phase terms in effect.

The data type for the delay LUTs (i.e. lut/delay/0 to lut/delay/3) is Fix_18_4.
This allows one to set the "constant" delays (i.e. delay0 to delay3) to a
midrange value and then use the delay LUTs to go plus or minus thar midrange
value.

The data type for the phase LUTs (i.e. lut/phase/0 to lut/phase/3) is
UFix_12_0, same as the "constant" phases (i.e. phase0 to phase3).

The LUTs are addressed by a counter that increments at the end of each complete
0/180 cycle.  The counter can be reset by the next walsh sync signal (1 PPS for
CARMA or SOWF for SMA, selected via walsh/syncsel).  This is done by a 0 to 1
transition on the LSb of lut/arm.  The current value of the LUT address counter
can be read from lut/addr.  Note that the LUT address counter counts even when
the LSb of lut/enable is 0.