Changed the SMA 0-180 Walsh cycle time in the IPA.
Also, added HB/SOWF signals to XAUI1 OOB in anticipation of sending these
signals to the DBE for use in the single-baseline comparison correlator.

This commit will be tagged as "ibob_phased_array_2k_2011_Mar_13_1744" and
was compiled using tag vlib-devel-2011-02-16 of the ral/mlib_devel library and
tag cfa_library-2011-02-16 of cfa_library.