Converted the DBE correlator to be complex.
Also added the ability to manually control delay and phase of the
analog input on the DBE, which should be the comparison antenna for
VLBI.

Another feature that has been built in but not tested is the removal
of an arbitrary artificial fringe rate. This is controlled by the
'comp/fringe_rate' register.