Add loadlut command to IPA TinySH
Also changed "break" statements to "return" statements in bramload_cmd to avoid
proceeding with bogus parameters.

The loadlut command is used to load a linear sequence of values into the
specified address of a BRAM look-up table (LUT).

Usage: loadlut <bram name> <address> <value> <step> <count>

The address will wrap around within the BRAM so it is safe to write past the
end of the BRAM.