Modify ibob_phased_array_2k_2011_Mar_31_1415.mcs
This version includes the following fixes and new feature:

* Fixed a cross-wiring between LUTs 0 and 1 (for both the delay and phase
  LUTs).

* Fixed a bug introduced by a "fix" in commit 4a43504.

* New tinysh command "loadlut" to load linear sequence into BRAMs (i.e. LUTs).
  This now supports the shift parameter.

This commit is what was used for the first VLBI recording of the CARMA phased
array!

This commit will be tagged as "ibob_phased_array_2k_2011_Mar_31_1415".
Note: this was compiled with tag vlbi-devel-2011-02-16 of the RAL fork of the
mlib_devel library (uses LVDS_25_DT for XAUI clocks) and tag
cfa_library-2011-02-16 of cfa_library.

Note that this commit supercedes commit 334a89c, which had superceded commit
60af604.  Both 334a89c and 60af604 had been tagged with
"ibob_phased_array_2k_2011_Mar_31_1415" at some point.