Extend all IPA->CC OOB syncs by one cycle
The start_xaui software register on ibob_phased_array_2k controls a one bit
counter that determines when the OOB signals to the calibration correlator are
captured, effectively downsampling the OOB signals by two.  If that counter
happens to be started on the cycle after the one cycle wide int_1pps or
int_1024pps pulses, those critical sync pulses would never get through to the
calibration correlator.  To address this, these one cycle wide pulses are
extended by one pulse so that they will be captured by the 1 bit counter
regardless of the software controlled startup phase of that counter.