Fix IPA->CC OOB/data downsample/demux logic
Prior to commit 21d54d8 there was a 50% chance that the 1 PPS and 1024 PPS sync
signals would not get through to the calibration correlator since the phase of
the OOB downsample-by-two was determined by the rising edge of a software
register (start_xaui).

Commit 21d54d8 attempted to solve this by extending 1 PPS and 1024 PPS to be
two cycles long.  While this did solve the 50% chance of missing sync pulses,
it created a 50% chance for the sync pulses to be delayed by on data sample.

To solve this, the phase of the demux-by-two is now determined by the 1024 PPS
cycle.  The start_xaui software register is now used to gate the demux-by-two
control signal so that the XAUI valid port is only driven high every other
cycle when start_xaui is high.