Add ibob_phased_array_2k_2011_Nov_12_1046.mcs
This version includes a fix to the IPA->CC OOB/data downsample/demux logic. See
commit c66e35d for details.  Commit 7eaac76 fixed timing issues of c66e35d.

This commit will be tagged as "ibob_phased_array_2k_2011_Nov_12_1046".
Note: this was compiled with tag vlbi-devel-2011-02-16 of the RAL fork of the
mlib_devel library (uses LVDS_25_DT for XAUI clocks) and tag
cfa_library-2011-02-16 of cfa_library.