Add ibob_phased_array_2k_2011_Nov_15_2206.mcs
This bitstream contains a fix for a bug that was introduced in commit 7eaac76
(bitstream added in commit 1237624).  That commit fixed a bug where sometimes
sync pulses from an IPA would not be received by the calibration correlator
(restarting the IPA's XAUI links would eventually fix the problem), but in
fixing that bug it introduced a new bug that caused a one cycle ambiguity in
the sync pulse received by the calibration correlator.

This commit will be tagged as "ibob_phased_array_2k_2011_Nov_15_2206".
Note: this was compiled with tag vlbi-devel-2011-02-16 of the RAL fork of the
mlib_devel library (uses LVDS_25_DT for XAUI clocks) and tag
cfa_library-2011-03-17 of cfa_library.