Enhance correlation options in dbe_xui_corr
The single baseline correlator can now be configured to correlate a wider
variery of input signal pairings than before.

The "comparison input" can be either the ADC or the digital noise source
(selected via the new "compsel" register).  The comparison input is passed
through delay_core.  This means that the noise source in the DBE can now be
correlated with noise sources in the IPAs, provided that the delays are set
properly.

The "insel" register now selects the signal that is sent to the data recorder
and the first input of the single baseline correlator.  The new "corrsel"
register selects the data that is sent to the snap block and the second input
of the single baseline correlator.  Both the insel and corrsel muxes select
from the same set of input signals:

0 = Final sum
1 = Comparison input (see compsel register above)
2 = Partial sum from XAUI 0
3 = Partial sum from XAUI 1