Removed trig_capt register
The trig_capt register has been removed from the model.  The trig "port" of the
snap block is now driven by "corr_sync" and the "we" port is hard-wired to 1.
This matches the snap block setup in bee2_complex_corr, thereby maintaining
more consistency across the different designs that make up the VLBI Phased
Array Processor.