Fix phase of data_valid vs 2-bit data demux-by-two
Commit d2d825c deprecated the start_xaui register.  The motivation for that was
to make the data_valid signal into the XAUI blocks determinitically related to
the demux-by-two.  That worked well enough except that the data_valid signal
needs to be deterministically related to the demux-by-two *and* be in
phase with the demux-by-two.  This commit fixes the the phase relationship
between the data_valid signal and the demux-by-two.