More signal path routing options
Renamed insel to vsisel.  Renamed corrsel to corrsel0.  Added corrsel1.

vsisel selects the input to the data recorder.
corrsel0 selects the first input to the single baseline correlator.
corrsel1 selects the second input to the single baseline correlator.

The first input to the correlator is also the input to the snap block.

This added mux allows correlating any two internal signals while recording a
third.  For exmaple, it allows the recording of the final sum while correlating
the two partial sums to verify their time alignment.

Also removed the old and now useless monitor2x16 subsystem.