Add details about the IPA rounding the 8-bit data
Added details about how shifting by 1 or two bits and then rounding to
even will affect the 8-bit sample distribution (more evens than odds).

Added info about the current gateware wrapping instead of saturating
when shifting the 8-bit sum by to bits.  This is arguably erroneous, but
the practical effects of this are negligible to none.