Add xaui_rst and valid registers to IPA design
The least significant bit of the "xaui_rst" register resets both XAUI yellow
blocks.  It also reset the 1-bit counter that generates the tx_valid signal for
the XAUI link to the DBE.  The tx_valid signal for the XAUI link to the CC is
reset every 1024 PPS.

The two top level subsystems, xaui0 and xaui1, each have a "valid" register
that captures the 32 most recent values of the XAUI yellow block's "tx_valid"
input signal.